requirements.txt
coreschema==0.0.4
Django==3.2.16
django-cors-headers==3.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.20.0
psycopg2>=2.8.6,<2.9
flake8>=3.9.2,<3.10
Pillow>=9.3,<9.4
sqlparse==0.3.1
urllib3==1.25.9
PyJWT==v1.7.1

